+++ src/main/java/si/matjazcerkvenik/alertmonitor/web/WebhookServlet.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package si.matjazcerkvenik.alertmonitor.webhook;
+package si.matjazcerkvenik.alertmonitor.web;
 
 import java.io.IOException;
 import java.util.Enumeration;